Get Gems improvement
